MPL-102 Fixed issue with path processing when running server on Win (#103)
This fix is for rare case of running Jenkins server on windows. The tokenization is not the great way of processing and expecting proper array items on the output was quite silly. So I rewrote the function to use File mechanism of parsing the paths - on *nix it will properly work with *nix paths and on windows it will properly process the windows paths.